Move gossiper and server to `bmmc` package
